fix: apply message color only if there is one
Sometimes you would like to use CSS instead of
JavaScript to apply color to the message bubble.
Not setting a style is impossible so far. But
if the color values are set to some empty
text, funny side effects may occur. So the value
are checked prior to applying them.