feature: introduce loading spinner
Loading of messages usually needs some time. Therefor the
text that no message are available yet will be displayed
at the beginning. Depending on the speed of the connection
to the server, it may just flicker, because loading is
fast enough. This is enoying to the user.

Now a loading spinner is displayed as long as the messages
are fetched from some server. Then - only of there are none -
the text indicating the absence of messages, will be displayed.

A new flag "isLoadingMessages" and a new template is being
introduced by this commit.